pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdm-egg"
version = "0.1.0"
description = "A minimal, CPU-friendly incubator for reinforcement learning on language models."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "Google DeepMind"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "jax>=0.4.0",
    "flax>=0.8.0",
    "optax>=0.2.0",
    "pandas",
    "numpy",
    "absl-py",
    "fancyflags",
]

[project.urls]
Homepage = "https://github.com/google-deepmind/egg"
Repository = "https://github.com/google-deepmind/egg"

[tool.setuptools.packages.find]
include = ["egg*", "experiments*"]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["experiments"]
